requirements.txt
streamlit
tensorflow
numpy
Pillow
scipy
matplotlib